Make agent install dependent on signed cert
Assume the signed cert will be retrieved from the master prior to
installing the new key. The reason for this requirement is that if a
cert doesn't exist, an agent will automatically request one from the
master. If the master gives it a cert signed by the old CA, the agent is
effectively disabled and cannot resume until manual intervention removes
the conflicting key and/or certificate. Therefore, don't install the key
unless the certificate is also available and can be installed at the
same time.